requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
torch
torchvision
pillow
clip-anytorch
vaderSentiment
scikit-learn